+++ nequip/scripts/evaluate.py
@@ -1,5 +1,6 @@
 import sys
 import argparse
+import logging
 import textwrap
 from pathlib import Path
 import contextlib
@@ -17,7 +18,7 @@ from nequip.train.loss import Loss
 from nequip.train.metrics import Metrics
 
 
-def main(args=None):
+def main(args=None, running_as_script: bool = True):
     # in results dir, do: nequip-deploy build . deployed.pth
     parser = argparse.ArgumentParser(
         description=textwrap.dedent(
@@ -25,7 +26,7 @@ def main(args=None):
 
             The model, metrics, dataset, etc. can specified individually, or a training session can be indicated with `--train-dir`.
 
-            Prints only the final result in `name = num` format to stdout; all other information is printed to stderr.
+            Prints only the final result in `name = num` format to stdout; all other information is logging.debuged to stderr.
 
             WARNING: Please note that results of CUDA models are rarely exactly reproducible, and that even CPU models can be nondeterministic.
             """
@@ -51,7 +52,7 @@ def main(args=None):
     )
     parser.add_argument(
         "--metrics-config",
-        help="A YAML config file specifying the metrics to compute. If omitted, `config_final.yaml` in `train_dir` will be used. If the config does not specify `metrics_components`, the default is to print MAEs and RMSEs for all fields given in the loss function. If the literal string `None`, no metrics will be computed.",
+        help="A YAML config file specifying the metrics to compute. If omitted, `config_final.yaml` in `train_dir` will be used. If the config does not specify `metrics_components`, the default is to logging.debug MAEs and RMSEs for all fields given in the loss function. If the literal string `None`, no metrics will be computed.",
         type=str,
         default=None,
     )
@@ -79,8 +80,14 @@ def main(args=None):
         type=Path,
         default=None,
     )
+    parser.add_argument(
+        "--log",
+        help="log file to store all the metrics and screen logging.debug",
+        type=Path,
+        default=None,
+    )
     # Something has to be provided
-    # See https://stackoverflow.com/questions/22368458/how-to-make-argparse-print-usage-when-no-option-is-given-to-the-code
+    # See https://stackoverflow.com/questions/22368458/how-to-make-argparse-logging.debug-usage-when-no-option-is-given-to-the-code
     if len(sys.argv) == 1:
         parser.print_help()
         parser.exit()
@@ -129,35 +136,60 @@ def main(args=None):
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
     else:
         device = torch.device(args.device)
-    print(f"Using device: {device}", file=sys.stderr)
+
+    if running_as_script:
+        # Configure the root logger so stuff gets printed
+        root_logger = logging.getLogger()
+        root_logger.setLevel(logging.CRITICAL)
+        root_logger.handlers = [
+            logging.StreamHandler(sys.stderr),
+            logging.StreamHandler(sys.stdout),
+        ]
+        root_logger.handlers[0].setLevel(logging.INFO)
+        root_logger.handlers[1].setLevel(logging.CRITICAL)
+        if args.log is not None:
+            root_logger.addHandler(logging.FileHandler(args.log, mode="w"))
+            root_logger.handlers[-1].setLevel(logging.INFO)
+    logger = logging.getLogger("nequip-evaluate")
+    logger.setLevel(logging.INFO)
+
+    logger.info(f"Using device: {device}")
     if device.type == "cuda":
-        print(
+        logger.info(
             "WARNING: please note that models running on CUDA are usually nondeterministc and that this manifests in the final test errors; for a _more_ deterministic result, please use `--device cpu`",
-            file=sys.stderr,
         )
 
     # Load model:
-    print("Loading model... ", file=sys.stderr, end="")
+    logger.info("Loading model... ")
     try:
         model, _ = load_deployed_model(args.model, device=device)
-        print("loaded deployed model.", file=sys.stderr)
+        logger.info("loaded deployed model.")
     except ValueError:  # its not a deployed model
         model = torch.load(args.model, map_location=device)
         model = model.to(device)
-        print("loaded pickled Python model.", file=sys.stderr)
+        logger.info("loaded pickled Python model.")
 
     # Load a config file
-    print(
-        f"Loading {'original training ' if dataset_is_from_training else ''}dataset...",
-        file=sys.stderr,
+    logger.info(
+        f"Loading {'original ' if dataset_is_from_training else ''}dataset...",
     )
     config = Config.from_file(str(args.dataset_config))
 
+    dataset_is_validation: bool = False
     # Currently, pytorch_geometric prints some status messages to stdout while loading the dataset
     # TODO: fix may come soon: https://github.com/rusty1s/pytorch_geometric/pull/2950
     # Until it does, just redirect them.
     with contextlib.redirect_stdout(sys.stderr):
-        dataset = dataset_from_config(config)
+        try:
+            # Try to get validation dataset
+            dataset = dataset_from_config(config, prefix="validation_dataset")
+            dataset_is_validation = True
+        except KeyError:
+            # Get shared train + validation dataset
+            dataset = dataset_from_config(config)
+    logger.info(
+        f"Loaded {'validation_' if dataset_is_validation else ''}dataset specified in {args.dataset_config.name}.",
+    )
 
     c = Collater.for_dataset(dataset, exclude_keys=[])
 
@@ -167,17 +199,22 @@ def main(args=None):
         # we know the train and val, get the rest
         all_idcs = set(range(len(dataset)))
         # set operations
-        test_idcs = list(all_idcs - train_idcs - val_idcs)
-        assert set(test_idcs).isdisjoint(train_idcs)
+        if dataset_is_validation:
+            test_idcs = list(all_idcs - val_idcs)
+            logger.info(
+                f"Using origial validation dataset minus validation set frames, yielding a test set size of {len(test_idcs)} frames.",
+            )
+        else:
+            test_idcs = list(all_idcs - train_idcs - val_idcs)
+            assert set(test_idcs).isdisjoint(train_idcs)
+            logger.info(
+                f"Using origial training dataset minus training and validation frames, yielding a test set size of {len(test_idcs)} frames.",
+            )
+        # No matter what it should be disjoint from validation:
         assert set(test_idcs).isdisjoint(val_idcs)
-        print(
-            f"Using training dataset minus training and validation frames, yielding a test set size of {len(test_idcs)} frames.",
-            file=sys.stderr,
-        )
         if not do_metrics:
-            print(
+            logger.info(
                 "WARNING: using the automatic test set ^^^ but not computing metrics, is this really what you wanted to do?",
-                file=sys.stderr,
             )
     else:
         # load from file
@@ -187,9 +224,8 @@ def main(args=None):
             ),
             filename=str(args.test_indexes),
         )
-        print(
+        logger.info(
             f"Using provided test set indexes, yielding a test set size of {len(test_idcs)} frames.",
-            file=sys.stderr,
         )
 
     # Figure out what metrics we're actually computing
@@ -224,7 +260,7 @@ def main(args=None):
     batch_i: int = 0
     batch_size: int = args.batch_size
 
-    print("Starting...", file=sys.stderr)
+    logger.info("Starting...")
     context_stack = contextlib.ExitStack()
     with contextlib.ExitStack() as context_stack:
         # "None" checks if in a TTY and disables if not
@@ -284,9 +320,8 @@ def main(args=None):
             display_bar.close()
 
     if do_metrics:
-        print(file=sys.stderr)
-        print("--- Final result: ---", file=sys.stderr)
-        print(
+        logger.info("\n--- Final result: ---")
+        logger.critical(
             "\n".join(
                 f"{k:>20s} = {v:< 20f}"
                 for k, v in metrics.flatten_metrics(metrics.current_result())[0].items()
@@ -295,4 +330,4 @@ def main(args=None):
 
 
 if __name__ == "__main__":
-    main()
+    main(running_as_script=True)