+++ CHANGELOG.md
@@ -10,6 +10,12 @@ Most recent change on the bottom.
 ### Added
 - Support for `e3nn`'s `soft_one_hot_linspace` as radial bases
 - Support for parallel dataloader workers with `dataloader_num_workers`
+- Optionally independently configure validation and training datasets
+- Save dataset parameters along with processed data
+
+### Changed
+- Name processed datasets based on a hash of their parameters to ensure only valid cached data is used
+- Do not use TensorFloat32 by default on Ampere GPUs until we understand it better
 
 ## [0.3.3] - 2021-08-11
 ### Added